* EquationSize cant be bigger than unsigned int (4 bytes so the maximum is 4'294'967'295) because of GUI realization. TODO: need to update containers size usage. * Error output small cleanup.
git-svn-id: https://studio2000.xyz:8443/svn/TPLTM/Lab1/Calc@415 bea4479c-5f86-c742-a837-c53f34f90b20

+++ Calc/flags.h
@@ -7,7 +7,7 @@
 /*
  * Type using as a type for calculation.
  */
-using EquationSize = size_t;
+using EquationSize = uint32_t; // TODO needs to rewrite usage in the containers.
 using Value = double;
 using ParamCount = char;
 //---------------------------------------------------------------------------
